+++ tests/conftest.py
@@ -1,0 +1,5 @@
+from .helper import call_command
+
+def pytest_sessionstart(session):
+    call_command("./cert/gencert.sh")
+
